requirements.txt
ollama
pytesseract
pdf2image
pillow
yagmail
python-dotenv
